Initial code cleaning
- modified air configs to fail on error
- restructured controller code
- updated formatting for template code
