+++ src/main/go/api/rest/common.go
@@ -1,0 +1,10 @@
+package rest
+
+const (
+	portfolioPath              = "/api/portfolio"
+	portfolioIdParam           = "portfolioId"
+	specificPortfolioPath      = portfolioPath + "/:" + portfolioIdParam
+	timeFrameTagParam          = "timeFrameTag"
+	planIdParam                = "planId"
+	getPortfolioIdErrorMessage = "Error getting portfolioId url parameter"
+)
