+++ src/main/go/api/rest/portfolio_rest_controller.go
@@ -2,8 +2,6 @@ package rest
 
 import (
 	"github.com/benizzio/open-asset-allocator/api/rest/model"
-	"github.com/benizzio/open-asset-allocator/application"
-	"github.com/benizzio/open-asset-allocator/domain"
 	"github.com/benizzio/open-asset-allocator/domain/service"
 	"github.com/benizzio/open-asset-allocator/infra"
 	"github.com/gin-gonic/gin"
@@ -11,19 +9,8 @@ import (
 	"strconv"
 )
 
-const (
-	portfolioPath              = "/api/portfolio"
-	portfolioIdParam           = "portfolioId"
-	specificPortfolioPath      = portfolioPath + "/:" + portfolioIdParam
-	timeFrameTagParam          = "timeFrameTag"
-	planIdParam                = "planId"
-	getPortfolioIdErrorMessage = "Error getting portfolioId url parameter"
-)
-
 type PortfolioRESTController struct {
-	portfolioDomService                 *service.PortfolioDomService
-	portfiolioAnalysisConfigService     *application.PortfolioAnalysisConfigurationAppService
-	portfiolioDivergenceAnalysisService *application.PortfolioDivergenceAnalysisAppService
+	portfolioDomService *service.PortfolioDomService
 }
 
 func (controller *PortfolioRESTController) BuildRoutes() []infra.RESTRoute {
@@ -42,16 +29,6 @@ func (controller *PortfolioRESTController) BuildRoutes() []infra.RESTRoute {
 			Method:   http.MethodGet,
 			Path:     specificPortfolioPath + "/history",
 			Handlers: gin.HandlersChain{controller.getPortfolioAllocationHistory},
-		},
-		{
-			Method:   http.MethodGet,
-			Path:     specificPortfolioPath + "/divergence/options",
-			Handlers: gin.HandlersChain{controller.GetDivergenceAnalysisOptions},
-		},
-		{
-			Method:   http.MethodGet,
-			Path:     specificPortfolioPath + "/divergence/:" + timeFrameTagParam + "/allocation-plan/:" + planIdParam,
-			Handlers: gin.HandlersChain{controller.GetDivergenceAnalysis},
 		},
 	}
 }
@@ -104,62 +81,8 @@ func (controller *PortfolioRESTController) getPortfolioAllocationHistory(context
 	context.JSON(http.StatusOK, aggregatedPortfoliohistoryDTS)
 }
 
-func (controller *PortfolioRESTController) GetDivergenceAnalysisOptions(context *gin.Context) {
-
-	portfolioIdParam := context.Param(portfolioIdParam)
-	portfolioId, err := strconv.Atoi(portfolioIdParam)
-	if infra.HandleAPIError(context, getPortfolioIdErrorMessage, err) {
-		return
-	}
-
-	analysisOptions, err := controller.portfiolioAnalysisConfigService.GetDivergenceAnalysisOptions(portfolioId)
-	if infra.HandleAPIError(context, "Error getting divergence analysis options", err) {
-		return
-	}
-
-	var analysisOptionsDTS = model.MapAnalysisOptions(analysisOptions)
-
-	context.JSON(http.StatusOK, analysisOptionsDTS)
-}
-
-func (controller *PortfolioRESTController) GetDivergenceAnalysis(context *gin.Context) {
-
-	portfolioIdParam := context.Param(portfolioIdParam)
-	portfolioId, err := strconv.Atoi(portfolioIdParam)
-	if infra.HandleAPIError(context, getPortfolioIdErrorMessage, err) {
-		return
-	}
-
-	var timeFrameTagParam = domain.TimeFrameTag(context.Param(timeFrameTagParam))
-
-	planIdParam := context.Param(planIdParam)
-	planId, err := strconv.Atoi(planIdParam)
-	if infra.HandleAPIError(context, "Error getting planId url parameter", err) {
-		return
-	}
-
-	analysis, err := controller.portfiolioDivergenceAnalysisService.GeneratePortfolioDivergenceAnalysis(
-		portfolioId,
-		timeFrameTagParam,
-		planId,
-	)
-	if infra.HandleAPIError(context, "Error generating portfolio divergence analysis", err) {
-		return
-	}
-
-	var analysisDTS = model.MapDivergenceAnalysis(analysis)
-
-	context.JSON(http.StatusOK, analysisDTS)
-}
-
-func BuildPortfolioRESTController(
-	portfolioDomService *service.PortfolioDomService,
-	portfiolioAnalysisConfigService *application.PortfolioAnalysisConfigurationAppService,
-	portfiolioAnalysisService *application.PortfolioDivergenceAnalysisAppService,
-) *PortfolioRESTController {
+func BuildPortfolioRESTController(portfolioDomService *service.PortfolioDomService) *PortfolioRESTController {
 	return &PortfolioRESTController{
 		portfolioDomService,
-		portfiolioAnalysisConfigService,
-		portfiolioAnalysisService,
 	}
 }
