+++ src/main/web-static/websrc/pages/portfolios.html
@@ -1,0 +1,14 @@
+<div id="portfolios"
+     hx-get="/api/portfolio"
+     hx-trigger="route-reached once"
+     data-hx-trigger-on-route="/*!route-reached"
+     handlebars-template="portfolios-management"
+     class="row justify-content-center align-items-center row-cols-auto my-5"
+></div>
+
+<script id="portfolio-data" type="text/x-handlebars-template">
+    <!--
+        TODO: cards for portfolio selection and form for portfolio creation
+        New portfolio should be a card on the end of the list that activates the form
+     -->
+</script>
